feat: Refactor draw graph func
