+++ src/components/organisms/CompanyOrderTable/index.tsx
@@ -133,12 +133,7 @@ const CompanyOrderTable = () => {
               </StyledTableCell>
               <StyledTableCell>{row.count}</StyledTableCell>
               <StyledTableCell>{row.price}</StyledTableCell>
-              {row.id !== 20 ? (
-                <StyledTableCell>{row.delivery_number === null ? '배송 전' : '배송 완료'}</StyledTableCell>
-              ) : null}
-              {row.id === 20 ? (
-                <StyledTableCell>{row.delivery_number === null ? '배송 전' : '배송 중'}</StyledTableCell>
-              ) : null}
+              <StyledTableCell>{row.delivery_number === null ? '배송 전' : '배송 완료'}</StyledTableCell>
               <StyledTableCell>
                 {row.delivery_number ? (
                   <a
