+++ src/components/template/MainTemplate/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { useHistory } from 'react-router-dom';
 import { Table, Container, Input, Button } from 'reactstrap';
-import { Dummy, date2String, calculateFee, makeMoneyStr } from 'utils';
+import { Dummy, date2String, calculateFee, makeMoneyStr, drawDoughnutGraph, drawLineGraph } from 'utils';
 import { Fade } from 'react-awesome-reveal';
 import { DoughnutChart, LineChart } from 'components/molecules';
 import { CompanyAdjustForm } from 'components/organisms';
@@ -46,15 +46,21 @@ const MainTemplate: React.FC = () => {
   const [totalFee, setTotalFee] = useState(0);
   const [lineGraph, setLineGraph] = useState(Dummy.chartData);
   const [doughnutGraph, setDoughnutGraph] = useState(Dummy.doughnutChartData);
-
+  const [isAdmin, setIsAdmin] = useState('' as string | null);
   const [countOrders, setCountOrders] = useState(0);
 
   useEffect(() => {
-    getAdjust();
-    getAdjustProducts();
-    getCompanyOrder();
-    getCompanyInfo();
+    setIsAdmin(localStorage.getItem('isAdmin'));
   }, []);
+
+  useEffect(() => {
+    if (isAdmin !== 'true') {
+      getAdjust();
+      getAdjustProducts();
+      getCompanyOrder();
+      getCompanyInfo();
+    }
+  }, [isAdmin]);
 
   useEffect(() => {
     calculateProfit();
@@ -62,11 +68,11 @@ const MainTemplate: React.FC = () => {
 
   useEffect(() => {
     setTotalFee(calculateFee(totalProfit));
-    drawDoughnutGraph();
+    setDoughnutGraph(drawDoughnutGraph(productProfitList, totalProfit));
   }, [totalProfit]);
 
   useEffect(() => {
-    drawLineGraph();
+    setLineGraph(drawLineGraph(adjustList));
   }, [adjustList]);
 
   const getAdjust = async () => {
@@ -110,92 +116,6 @@ const MainTemplate: React.FC = () => {
       return 0;
     });
     setTotalProfit(total);
-  };
-
-  const drawLineGraph = () => {
-    const lineLabels = ['', '', '', '', '', ''];
-    const lineData = [0, 0, 0, 0, 0, 0];
-    let monthAgo = new Date();
-    monthAgo = new Date(monthAgo.getFullYear(), monthAgo.getMonth(), 1);
-
-    for (let i = 1; i <= 6; i += 1) {
-      monthAgo = new Date(monthAgo.getFullYear(), monthAgo.getMonth() - 1, 1);
-      lineLabels[6 - i] = date2String(monthAgo);
-    }
-
-    adjustList.map(ad => {
-      const date = ad.date.substr(0, 7);
-      const idx = lineLabels.findIndex(el => date === el);
-      if (idx !== -1) lineData[idx] = ad.price;
-      return 0;
-    });
-    const graphData = {
-      labels: lineLabels,
-      datasets: [
-        {
-          label: '월별 수익',
-          data: lineData,
-          fill: false,
-          backgroundColor: 'rgb(97, 157, 160)',
-          borderColor: 'rgba(97, 157, 160, 0.5)',
-        },
-      ],
-    };
-    setLineGraph(graphData);
-  };
-
-  const drawDoughnutGraph = () => {
-    const doughnutLabels = [] as Array<string>;
-    const doughnutData = [] as Array<number>;
-
-    if (productProfitList.length === 0) {
-      doughnutLabels.push('판매된 상품이 없습니다');
-      doughnutData.push(100);
-    }
-    const isMany = productProfitList.length > 6;
-    let total = 0;
-    const tempList = productProfitList.sort((a, b) => Number(b.total_price) - Number(a.total_price));
-    tempList.some((pro, idx) => {
-      if (idx === 6) return true;
-      if (idx === 5 && isMany) {
-        doughnutData.push(100 - total);
-        doughnutLabels.push('기타');
-      }
-      let price = 0;
-      if (totalProfit !== 0) price = (Number(pro.total_price) / totalProfit) * 100;
-      total += price;
-      doughnutData.push(price);
-      doughnutLabels.push(pro.name);
-
-      return false;
-    });
-    const graphData = {
-      labels: doughnutLabels,
-      datasets: [
-        {
-          label: '상품 판매량',
-          data: doughnutData,
-          backgroundColor: [
-            'rgba(255, 99, 132, 0.2)',
-            'rgba(54, 162, 235, 0.2)',
-            'rgba(255, 206, 86, 0.2)',
-            'rgba(75, 192, 192, 0.2)',
-            'rgba(153, 102, 255, 0.2)',
-            'rgba(255, 159, 64, 0.2)',
-          ],
-          borderColor: [
-            'rgba(255, 99, 132, 1)',
-            'rgba(54, 162, 235, 1)',
-            'rgba(255, 206, 86, 1)',
-            'rgba(75, 192, 192, 1)',
-            'rgba(153, 102, 255, 1)',
-            'rgba(255, 159, 64, 1)',
-          ],
-          borderWidth: 1,
-        },
-      ],
-    };
-    setDoughnutGraph(graphData);
   };
 
   const useStyles = makeStyles({
